feat: adding camera video list
